Added protection (session requirement) for some api routes Removed redundant information from JWT token and session

+++ src/pages/api/form/index.js
@@ -1,7 +1,16 @@
 import prisma from "@/server";
+import {getServerSession} from "next-auth";
+import {authOptions} from "@/pages/api/auth/[...nextauth]";
 
 export default async function handler(req, res) {
     if (req.method === 'POST') {
+        const session = await getServerSession(req, res, authOptions)
+
+        if (!session) {
+            res.status(401).json({ message: "You must be logged in." });
+            return;
+        }
+
         const { body } = req
         const { userId } = body
         const form = await prisma.form.create({
@@ -36,6 +45,13 @@ export default async function handler(req, res) {
         }
         res.send({form})
     } else if(req.method === 'PATCH') {
+        const session = await getServerSession(req, res, authOptions)
+
+        if (!session) {
+            res.status(401).json({ message: "You must be logged in." });
+            return;
+        }
+
         const { body } = req
         const { id, active, description, name, questions } = body
 
@@ -78,6 +94,13 @@ export default async function handler(req, res) {
         }
         res.status(200).send(form)
     } else if (req.method === 'DELETE'){
+        const session = await getServerSession(req, res, authOptions)
+
+        if (!session) {
+            res.status(401).json({ message: "You must be logged in." });
+            return;
+        }
+
         const {query} = req
         const {id} = query
         await prisma.form.delete({
